google-cloud/b24-spa-1106-sync/b24-spa-1106-daily-sync/requirements.txt
functions-framework==3.*
google-cloud-pubsub==2.*
google-cloud-secret-manager==2.*
requests==2.*
flask==3.*
